pyproject.toml
[project]
name = "flow-builder-backend"
version = "0.1.0"
description = "Backend API for the Flow Builder application."
license = { file = "LICENCE" }
readme = { file = "README.md", content-type = "text/markdown" }
requires-python = ">=3.12"
authors = [{ name = "Chitrank Agnihoti", email = "chitrank2050@gmail.com" }]

dependencies = [
  "fastapi>=0.123.8",
  "loguru>=0.7.3",
  "pydantic-settings>=2.13.1",
  "python-dotenv>=1.2.2",
  "uvicorn>=0.41.0",
]

[dependency-groups]
dev = ["build>=1.4.0", "ruff>=0.14.13"]

[project.scripts]
dev = "app.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
